Restructuring of test folder
