+++ tests/bofire/data_models/constraints/test_nchoosek_combinatorics.py
