+++ tests/bofire/data_models/domain/test_numeric.py
